Renamed Integrity to IntegrityMAC, used MessageDigests for Integrity
IntegrityMAC provides integrity plus authentication
Integrity just provides integrity
